Deprecate cr index "charts-repo" flag. Instead read index.yaml from git repository (#144)
* read index.yaml from git repository

So far the index.yaml was loaded directly from the chart repository.
In case of GitHub pages that's https://<user or org>.github.io/<repo>

The problem with this approach is that GitHub pages use a Content
Delivery Network. So the index.yaml files fetched from there might be
outdated as there is a delay between a git push to the gh-pages branch
until the content is visible.

This delay might lead to the situation that one downloads an old index
file from and updated that one instead of the most recent index.yaml.

For the grafana repository that resulted in already released charts
being removed again.
- https://github.com/grafana/helm-charts/issues/783
- https://github.com/grafana/helm-charts/pull/796

This PR resolves that issue by reading the index.yaml from the gh_pages
branch of the repository. The branch needs to be up-to-date anyhow if
you want to create a PR or directly push the changes.

It should also be fine for the chart-releaser-action as it uses
a fetch-depth: 0 to clone the whole repository.

```
jobs:
  release:
    runs-on: ubuntu-latest
    steps:
      - name: Checkout
        uses: actions/checkout@v2
        with:
          fetch-depth: 0
```

Fixes: #143

Signed-off-by: Torsten Walter <torsten.walter@syncier.com>

* fix error handling

Signed-off-by: Torsten Walter <torsten.walter@syncier.com>

* re-add charts-repo flag to avoid breaking changes

- message is printed when using the flag to let users know that it's
  deprecated
- update all docs in README

Signed-off-by: Torsten Walter <torsten.walter@syncier.com>

* Update cr/cmd/index.go

Co-authored-by: Reinhard Nägele <unguiculus@gmail.com>
Signed-off-by: Torsten Walter <mail@torstenwalter.de>

* fix import ordering

Signed-off-by: Torsten Walter <torsten.walter@syncier.com>

Co-authored-by: Reinhard Nägele <unguiculus@gmail.com>